WIP Add embedded Python code
